python-services/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
moviepy==1.0.3
Pillow==10.2.0
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
